requirements.txt
selenium==4.15.0
pytest==8.0.0
pytest-html==4.1.1
allure-pytest==2.13.5
webdriver-manager==4.0.2
pytest-xdist==3.5.0
pytest-timeout==2.2.0